MB-33463: Refractor compaction_ctx to deal with implicit compactions
The current compaction context was using a unique_ptr and when the
explicit compaction completes, it resets the ptr. If an implicit
compaction started up, it could grab that ptr and then it gets freed
while the compaction was happening causing a race condition.

Solution is to change to shared pointer.

We are still waiting on the compaction callback to retrieve the
correct compaction context info. This is continuing the current
stopgap solution.

Change-Id: I38435085a7d82b16178b10eb8d68026eb4f6afcc