MB-33463: Add support for magma memory quota
Memory is maintained across 3 magma components; Bloom filters, Block
cache and Write cache. The least important of these is the write cache.
If there is insufficient memory for the write cache, the write cache
will grow to the size of the batch and then be immediately flushed and
freed. If there is available memory, the write cache is limited to 20%
of the available memory (after bloom filter and block cache get their
memory up to magma_max_write_cache (128MB). Bloom filters are the most
important and are never paged out. Bloom filter memory can cause magma
to go above the memory quota. To alleviate this, the bottom layer where
the majority of bloom filter memory is, won't use bloom filters when
OptimizeBloomFilterForMisses is on (which it is by default). The block
cache grows each time the index sizes change. But its growth is bounded
by the available memory or what's left over after the bloom filter
memory is subtracted.

Change-Id: I83edc9e7caa7515cccaa63fe27e0b39f4588e121